# docs/swagger.yaml
basePath: /api/v1
definitions:
  response.Msg:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:9999
info:
  contact: {}
  description: Mini ECS 的API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Mini ECS API
  version: "1.0"
paths:
  /course:
    get:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 获取课程列表
      tags:
      - course management
    post:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 创建课程
      tags:
      - course management
  /course/:uuid:
    delete:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 删除课程
      tags:
      - course management
    get:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 获取课程信息
      tags:
      - course management
    put:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 修改课程
      tags:
      - course management
  /image:
    get:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 获取镜像列表
      tags:
      - image management
    post:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 创建（上传）镜像
      tags:
      - image management
  /image/:uuid:
    delete:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 删除镜像
      tags:
      - image management
    get:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 获取镜像具体信息
      tags:
      - image management
    put:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 修改镜像条目信息
      tags:
      - image management
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 用户登录
      tags:
      - user
  /user/modify:
    post:
      consumes:
      - application/json
      description: 用户修改个人信息
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 用户修改个人信息
      tags:
      - user
  /user/register:
    post:
      consumes:
      - application/json
      description: 用户注册
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 用户注册
      tags:
      - user
  /vm:
    get:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 获取实例列表
      tags:
      - virtual machine management
    post:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 创建实例
      tags:
      - virtual machine management
  /vm/:uuid:
    delete:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 删除实例
      tags:
      - virtual machine management
    get:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 获取实例信息
      tags:
      - virtual machine management
    put:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: Unimplemented
      tags:
      - virtual machine management
  /vm/image:
    post:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 根据实例创建镜像
      tags:
      - virtual machine management
  /vm/snapshot:
    patch:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 将实例恢复到某个快照
      tags:
      - virtual machine management
    post:
      consumes:
      - application/json
      description: Unimplemented
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: passwd
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Msg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 根据实例创建快照
      tags:
      - virtual machine management
  /welcome:
    get:
      consumes:
      - application/json
      description: 操作行为的详细说明。
      parameters:
      - description: 这是评论
        in: query
        name: val1
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 这是评论1
          schema:
            type: integer
        "400":
          description: 这是评论2
          schema:
            $ref: '#/definitions/response.Msg'
      summary: 该操作的简短摘要
      tags:
      - example-tag
swagger: "2.0"