# internal/router/router.go
package router

import (
	"fmt"
	"github.com/gin-gonic/gin"
	"github.com/mini-ecs/back-end/api/v1"
	"github.com/mini-ecs/back-end/pkg/common/response"
	"github.com/mini-ecs/back-end/pkg/log"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
	"net/http"
	"strings"
)

var logger = log.GetGlobalLogger()

func NewRouter() *gin.Engine {
	gin.SetMode(gin.DebugMode)
	r := gin.Default()
	r.Use(Cors())
	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	group := r.Group("api/v1")
	{
		group.GET("/welcome", v1.Welcome)

		group.POST("/user/login", v1.Login)
		group.POST("/user/register", v1.RegisterUser)
		group.POST("/user/modify", v1.ModifyUser)

		group.GET("/course", v1.GetCourseList)
		group.GET("/course/:uuid", v1.GetSpecificCourse)
		group.PUT("/course/:uuid", v1.ModifyCourse)
		group.DELETE("/course/:uuid", v1.DeleteCourse)
		group.POST("/course", v1.CreateCourse)

		group.GET("/image", v1.GetImageList)
		group.GET("/image/:uuid", v1.GetSpecificImage)
		group.PUT("/image/:uuid", v1.ModifyImage)
		group.DELETE("/image/:uuid", v1.DeleteImage)
		group.POST("/image", v1.CreateImage)

		group.GET("/vm", v1.GetVMList)
		group.GET("/vm/:uuid", v1.GetSpecificVM)
		group.PUT("/vm/:uuid", v1.ModifyVM)
		group.DELETE("/vm/:uuid", v1.DeleteVM)
		group.POST("/vm", v1.CreateVM)
		group.POST("/vm/image", v1.MakeImageWithVM)
		group.POST("/vm/snapshot", v1.MakeSnapshotWithVM)
		group.PATCH("/vm/snapshot", v1.ResetVMWithSnapshot)

	}
	return r
}
func Recovery(c *gin.Context) {
	defer func() {
		if r := recover(); r != nil {
			logger.Errorf("gin catch error: %v", r)
			c.JSON(http.StatusOK, response.FailMsg("系统内部错误"))
		}
	}()
	c.Next()
}
func Cors() gin.HandlerFunc {
	return func(c *gin.Context) {
		method := c.Request.Method               //请求方法
		origin := c.Request.Header.Get("Origin") //请求头部
		var headerKeys []string                  // 声明请求头keys
		for k, _ := range c.Request.Header {
			headerKeys = append(headerKeys, k)
		}
		headerStr := strings.Join(headerKeys, ", ")
		if headerStr != "" {
			headerStr = fmt.Sprintf("access-control-allow-origin, access-control-allow-headers, %s", headerStr)
		} else {
			headerStr = "access-control-allow-origin, access-control-allow-headers"
		}
		if origin != "" {
			c.Writer.Header().Set("Access-Control-Allow-Origin", "*")
			c.Header("Access-Control-Allow-Origin", "*")                                       // 这是允许访问所有域
			c.Header("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE,UPDATE") //服务器支持的所有跨域请求的方法,为了避免浏览次请求的多次'预检'请求
			//  header的类型
			c.Header("Access-Control-Allow-Headers", "Authorization, Content-Length, X-CSRF-Token, Token,session,X_Requested_With,Accept, Origin, Host, Connection, Accept-Encoding, Accept-Language,DNT, X-CustomHeader, Keep-Alive, User-Agent, X-Requested-With, If-Modified-Since, Cache-Control, Content-Type, Pragma")
			//              允许跨域设置                                                                                                      可以返回其他子段
			c.Header("Access-Control-Expose-Headers", "Content-Length, Access-Control-Allow-Origin, Access-Control-Allow-Headers,Cache-Control,Content-Language,Content-Type,Expires,Last-Modified,Pragma,FooBar") // 跨域关键设置 让浏览器可以解析
			c.Header("Access-Control-Max-Age", "172800")                                                                                                                                                           // 缓存请求信息 单位为秒
			c.Header("Access-Control-Allow-Credentials", "false")                                                                                                                                                  //  跨域请求是否需要带cookie信息 默认设置为true
			c.Set("content-type", "application/json")                                                                                                                                                              // 设置返回格式是json
		}
		//放行所有OPTIONS方法
		if method == "OPTIONS" {
			c.JSON(http.StatusOK, "Options Request!")
		}
		// 处理请求
		c.Next() //  处理请求
	}
}